move code into src and fix imports relatively to src
