+++ server/src/api.py
@@ -4,11 +4,11 @@ from fastapi import APIRouter, Depends, FastAPI, File, Header, Request, UploadFi
 from fastapi.middleware.cors import CORSMiddleware
 from pydantic import BaseModel
 
-from .model import QueryItem
+from model import QueryItem
 
-from . import model
-from .config import logger
-from .sqlclient import client
+import model
+from config import logger
+from sqlclient import client
 
 app = FastAPI()
